GH-592 Move files back as they are used by block log splitting as well
